+++ tests/tests_token/mod.rs
@@ -1,3 +1,4 @@
 mod auth;
 mod crypto;
 mod csrf;
+mod security;
